moved enum to enum folder
