+++ beta-src/src/components/ui/WDPositionContainer.tsx
@@ -1,13 +1,7 @@
 import * as React from "react";
 import Box from "@mui/material/Box";
 import { useMemo } from "react";
-
-export enum Position {
-  BOTTOM_LEFT = "bottom-left",
-  BOTTOM_RIGHT = "bottom-right",
-  TOP_LEFT = "top-left",
-  TOP_RIGHT = "top-right",
-}
+import Position from "../../enums/Position";
 
 interface WDPositionContainerProps {
   children: React.ReactNode;
