+++ beta-src/src/enums/Position.ts
@@ -1,0 +1,8 @@
+export enum Position {
+  BOTTOM_LEFT = "bottom-left",
+  BOTTOM_RIGHT = "bottom-right",
+  TOP_LEFT = "top-left",
+  TOP_RIGHT = "top-right",
+}
+
+export default Position;
